+++ typescript/packages/core/src/index.ts
@@ -1,60 +1,10 @@
-import type { Plugin } from "./plugins/plugins";
-import { sendETH } from "./plugins/send-eth";
-import { sendSOL } from "./plugins/send-sol";
-import {
-    type DeferredTool,
-    type GetDeferredToolsParams,
-    type GetToolsParams,
-    type Tool,
-    getDeferredTools,
-    getTools,
-} from "./tools";
-import { addParametersToDescription, parametersToJsonExample } from "./utils";
-import type {
-    Balance,
-    Chain,
-    ChainForWalletClient,
-    EVMReadRequest,
-    EVMSmartWalletClient,
-    EVMTransaction,
-    EVMTypedData,
-    EVMWalletClient,
-    Signature,
-    SolanaReadRequest,
-    SolanaTransaction,
-    SolanaWalletClient,
-    WalletClient,
-    isEVMSmartWalletClient,
-    isEVMWalletClient,
-    isSolanaWalletClient,
-} from "./wallets";
-
-export {
-    getTools,
-    getDeferredTools,
-    sendETH,
-    sendSOL,
-    addParametersToDescription,
-    parametersToJsonExample,
-    type Tool,
-    type DeferredTool,
-    type GetToolsParams,
-    type GetDeferredToolsParams,
-    type Plugin,
-    type WalletClient,
-    type EVMTransaction,
-    type EVMReadRequest,
-    type EVMWalletClient,
-    type EVMSmartWalletClient,
-    type SolanaTransaction,
-    type SolanaReadRequest,
-    type SolanaWalletClient,
-    type Signature,
-    type Balance,
-    type EVMTypedData,
-    type isEVMWalletClient,
-    type isEVMSmartWalletClient,
-    type isSolanaWalletClient,
-    type Chain,
-    type ChainForWalletClient,
-};
+export * from "./plugins";
+export * from "./get-tools";
+export * from "./utils";
+export * from "./wallets";
+export * from "./tool";
+export * from "./core-plugins/send-chr";
+export * from "./core-plugins/send-eth";
+export * from "./core-plugins/send-sol";
+export * from "./core-plugins/evm-send-transaction";
+export * from "./core-plugins/send-apt";