+++ typescript/examples/vercel-ai/crossmint-smart-wallets/CHANGELOG.md
@@ -1,5 +1,61 @@
 # goat-examples-vercel-ai-crossmint-smart-wallets
 
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [a66ceec]
+  - @goat-sdk/adapter-vercel-ai@0.1.9
+  - @goat-sdk/core@0.3.15
+  - @goat-sdk/plugin-erc20@0.1.12
+  - @goat-sdk/crossmint@0.2.6
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [fad74d5]
+  - @goat-sdk/crossmint@0.2.3
+
+## 0.1.5
+
+### Patch Changes
+
+- 16271ab: Simplify core architecture
+- Updated dependencies [16271ab]
+  - @goat-sdk/crossmint@0.2.2
+  - @goat-sdk/plugin-erc20@0.1.11
+  - @goat-sdk/core@0.3.14
+  - @goat-sdk/adapter-vercel-ai@0.1.8
+
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [cad06c0]
+- Updated dependencies [ad45f94]
+  - @goat-sdk/core@0.3.13
+  - @goat-sdk/adapter-vercel-ai@0.1.7
+  - @goat-sdk/plugin-erc20@0.1.10
+  - @goat-sdk/crossmint@0.2.1
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [bf161b7]
+  - @goat-sdk/crossmint@0.2.0
+  - @goat-sdk/core@0.3.12
+  - @goat-sdk/adapter-vercel-ai@0.1.6
+  - @goat-sdk/plugin-erc20@0.1.9
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [c2dbb80]
+  - @goat-sdk/crossmint@0.1.7
+
 ## 0.1.1
 
 ### Patch Changes