+++ typescript/examples/vercel-ai/solana/CHANGELOG.md
@@ -1,5 +1,56 @@
 # goat-examples-vercel-ai-solana
 
+## 0.1.6
+
+### Patch Changes
+
+- a66ceec: Fix jupiter swap tokens tx
+- Updated dependencies [a66ceec]
+  - @goat-sdk/adapter-vercel-ai@0.1.9
+  - @goat-sdk/plugin-spl-token@0.0.5
+  - @goat-sdk/plugin-jupiter@0.0.4
+  - @goat-sdk/wallet-solana@0.1.9
+  - @goat-sdk/core@0.3.15
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [d289856]
+  - @goat-sdk/wallet-solana@0.1.8
+
+## 0.1.4
+
+### Patch Changes
+
+- 16271ab: Simplify core architecture
+- Updated dependencies [16271ab]
+  - @goat-sdk/plugin-erc20@0.1.11
+  - @goat-sdk/core@0.3.14
+  - @goat-sdk/adapter-vercel-ai@0.1.8
+  - @goat-sdk/wallet-solana@0.1.7
+
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [cad06c0]
+- Updated dependencies [ad45f94]
+  - @goat-sdk/core@0.3.13
+  - @goat-sdk/adapter-vercel-ai@0.1.7
+  - @goat-sdk/plugin-erc20@0.1.10
+  - @goat-sdk/wallet-solana@0.1.6
+
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [bf161b7]
+  - @goat-sdk/core@0.3.12
+  - @goat-sdk/adapter-vercel-ai@0.1.6
+  - @goat-sdk/plugin-erc20@0.1.9
+  - @goat-sdk/wallet-solana@0.1.5
+
 ## 0.1.1
 
 ### Patch Changes