+++ typescript/packages/adapters/vercel-ai/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @goat-sdk/adapter-vercel-ai
 
+## 0.1.9
+
+### Patch Changes
+
+- a66ceec: Fix jupiter swap tokens tx
+- Updated dependencies [a66ceec]
+  - @goat-sdk/core@0.3.15
+
+## 0.1.8
+
+### Patch Changes
+
+- 16271ab: Simplify core architecture
+- Updated dependencies [16271ab]
+  - @goat-sdk/core@0.3.14
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [cad06c0]
+- Updated dependencies [ad45f94]
+  - @goat-sdk/core@0.3.13
+
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [bf161b7]
+  - @goat-sdk/core@0.3.12
+
 ## 0.1.5
 
 ### Patch Changes