+++ typescript/packages/core/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @goat-sdk/core
 
+## 0.3.15
+
+### Patch Changes
+
+- a66ceec: Fix jupiter swap tokens tx
+
+## 0.3.14
+
+### Patch Changes
+
+- 16271ab: Simplify core architecture
+
+## 0.3.13
+
+### Patch Changes
+
+- cad06c0: uniswap package and example
+- ad45f94: Added Chromia and send-chr tools
+
+## 0.3.12
+
+### Patch Changes
+
+- bf161b7: Add mint plugin to Crossmint
+
 ## 0.3.11
 
 ### Patch Changes