+++ typescript/packages/plugins/polymarket/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @goat-sdk/plugin-polymarket
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [a66ceec]
+  - @goat-sdk/core@0.3.15
+
+## 0.2.4
+
+### Patch Changes
+
+- 16271ab: Simplify core architecture
+- Updated dependencies [16271ab]
+  - @goat-sdk/core@0.3.14
+
+## 0.2.3
+
+### Patch Changes
+
+- Updated dependencies [cad06c0]
+- Updated dependencies [ad45f94]
+  - @goat-sdk/core@0.3.13
+
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies [bf161b7]
+  - @goat-sdk/core@0.3.12
+
 ## 0.2.1
 
 ### Patch Changes