+++ typescript/packages/wallets/crossmint/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @goat-sdk/crossmint
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [a66ceec]
+  - @goat-sdk/core@0.3.15
+
+## 0.2.5
+
+### Patch Changes
+
+- fad74d5: Fix content type headers missing on api calls
+
+## 0.2.2
+
+### Patch Changes
+
+- 16271ab: Simplify core architecture
+- Updated dependencies [16271ab]
+  - @goat-sdk/core@0.3.14
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [cad06c0]
+- Updated dependencies [ad45f94]
+  - @goat-sdk/core@0.3.13
+
+## 0.2.0
+
+### Minor Changes
+
+- bf161b7: Add mint plugin to Crossmint
+
+### Patch Changes
+
+- Updated dependencies [bf161b7]
+  - @goat-sdk/core@0.3.12
+
+## 0.1.7
+
+### Patch Changes
+
+- c2dbb80: Fix locator of custodial wallets
+
 ## 0.1.6
 
 ### Patch Changes